Move releasing (update images) into the daemon
(although it's not accessible yet)

This involves a lot of chopping and changing, mainly because releases
are no longer jobs, and where components come from has changed; and,
there's no especial need for a releaser component.

Also: deprecate the previous Platform implementations, so that old
daemons will be able to connect but not do much more. (Consider
deprecating them entirely, so daemons can't connect). Follow through
in the rpc, bus etc.

Next: expose Release in the daemon's API.